Update patch set 3
Patch Set 3:

As per the discussion with Alec, a further more robust refinement can be:
   if (pd[port].timeout - now.val < timeout)
==>
   if ((int)(pd[port].timeout - now.val) < timeout)

Otherwise, the comparison of (uint64) - (uint64) < (int32) may lurk potential bug when negative outcome is resulted

Another way of change can be to declare timeout from int32 to uint32.

We may revisit this later.

Patch-set: 3