Update patch set 1
Patch Set 1:

(1 comment)

I'll put a caller up today and you can take a look, I could also add to this CL changes to the existing code that parse setup packets by hand if you think that would be useful, I didn't initially because it seemed like a bit much churn.  I'm happy to wait until the use of this CL goes up for you to reconsider it.

Patch-set: 1