Update patch set 2
Patch Set 2:

This problem was found when I was testing BizLink's board files on STM32F-Discovery EVA board (we are still figuring out the how-to of HW/FW verification of our hoho/dingdong/docking boards).

However, were to submit this CL, the only way I can perform in TEST now, is 'make buildall'. So, wonder what's better way to do the _TEST_ in the future, in particular, under real boards !!

Patch-set: 2
Label: Code-Review=0