Update patch set 2
Patch Set 2:

ww: is for write a word to memory
rw: is for read or write a word from/to memory
md: is for memory dump or read a word from memory

If we want to keep a min EC code size, we only need keep one CLI, rw.

If we want a user friendly CLI (like uboot), we can add a CLI like md.
 
http://www.denx.de/wiki/DULG/Manual
5.9.2.5. md	- memory display

Since I add md, "rw" can be removed.

Patch-set: 2