Makefile: remove burn_my_ec
burn_my_ec is an utility that flash an image embedded in its code.
We can not compile it as part of ec-[dev]utils, because we have
devices that firmware should be build as part of chrome-ec package.
Remove burn_my_ec, barely used.

Split the makefile to build just the host utility when requested.

BRANCH=ToT
BUG=chrome-os-partner:32025,chromium:408713
TEST=Check that files are stil built when needed and
not when utils-host is invoked.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/233347
(cherry picked from commit 0777ebaef6ea1017b897cde65b2a9d1a39570089)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: I3fabe16067d57c74ae36b05138f4c6fd2483c7c4