Update patch set 5
Patch Set 5:

I believe so. Currently I don't cache any of the power supply status values but I intend to do so in order to reduce traffic to the PD MCU.  If the command fails and we don't get a notification, we may be left in a confused state with corner cases to deal with.

Patch-set: 5
Reviewer: Gerrit User 1000030 <1000030@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0