samus: fix EC watchdog when servicing pd flashing commands
Fix EC watchdog when getting a lot of host commands. This happens
on a zinger update because of the samus back-boosting workaround
in which the hooks task tries to send an i2c message to the BQ
on AC connect, but that i2c message can't go out because the host
command task holds the mutex and doesn't give it up for a long
enough period of time.

BUG=chrome-os-partner:33905
BRANCH=samus
TEST=loaded onto samus, tested remote update of zinger 10 times
and succeeded 10 times, no reboots and RSA signature check passed.

Change-Id: I42b88fb2f1611d36d97211874f15bc66ccb89db3
Signed-off-by: Alec Berg <alecaberg@chromium.org>