Update patch set 2
Patch Set 2:

The 1% boot test passed, but I think it may be safer to leave the threshold at 2%. In my testing, 1% remaining will be reported, up until the point that remaining capacity is reported as 0 mAh! The system continues to function when 0mAh is reported. Inserting a charger for a short time brings us up to 24 mAh, and I succeeded to boot to OS, get a  battery warning logo flash, and auto-power-down.

I should probably investigate how battery remaining capacity is reported when we're nearly empty.

Patch-set: 2