Update patch set 1
Patch Set 1:

It's not about the debug commands on the serial console. It's about faft being able to send a keyboard command while the firmware polls for keyboard commands.

In this case faft uses the ec console as if it used the keyboard (through kbpress). We have the real keyboard task higher priority than this, so why wouldn't faft's kbpress also be high priority (especially given that in low priority faft cannot actually send a kbpress command due to missing uart characters).

Patch-set: 1