samus: when battery is full request 12V to save power
When battery if full and zinger or minimuffin is plugged in,
request 12V to save power. Once battery is not full anymore,
or zinger or minimuffin is not plugged in, then change default
request back up to 20V.

BUG=chrome-os-partner:34640
BRANCH=samus
TEST=load onto samus. use battfake command from pd console to
set battery percentage. when zinger is plugged in, see that
batt >= 99% sends new 12V request to zinger, and when batt <= 97%
sends new 20V request. when zinger is unplugged, see that the
default request goes back to 20V.

tested that we receive host command from EC with battery
information every time battery changes SOC.

also tested during a zinger firmware update, fake battery SOC
goes above 99% and we send a new request and verified that zinger
update is still successful

Change-Id: Ia8e0721508e34ee3630f5e5b0c2f431a00329caf
Signed-off-by: Alec Berg <alecaberg@chromium.org>