Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

The change seems ok in itself.  Would be nice to verify this won't do anything bad to other boards which use STM32F0.  (Are those now a *lot* slower, or just a little?  I know I2C doesn't really have a minimum speed, but if this change makes those boards run at 50kbps or something like that, we may need this to be configurable per-board.

Patch-set: 4