Update patch set 1
Patch Set 1:

I found https://chromium-review.googlesource.com/#/c/225771/, it looks like it fixes something like this.
Reverting it doesn't make the problem come back though.

Victor: When did this get fixed? Why is it not a problem anymore? It used to not work in Sept.

I guess I could do a bisect if we can't find an answer.

Patch-set: 1