samus: fix tap for battery doesn't work in G3 with AC
Fix bug causing tap for battery not to work in G3 when AC
is attached. Problem was that the lightbar was being held
in reset and would not light up.

BUG=chrome-os-partner:34722
BRANCH=samus
TEST=tested on samus with AC attached and unattached in S3
and in G3, tap for battery lights up the lightbar. Also
verified that plugging and unpluggin AC in S3 and G3 causes
lightbar to light up.

Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236560
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
(cherry picked from commit ee5dbe94a0896b3edadd539ca4a2682ba718a205)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: I0fbd989399d372d287467200512bbdf2551884d3