Update patch set 2
Patch Set 2:

Still has one bug: if zinger is plugged in while battery is full, we don't change to 12V because when board_check_vbus_request() is run on connect, we don't know it's a zinger. when we get the response to discover identity, we don't call board_check_vbus_request() again. and since battery is full, the SOC % will never change, so we never re-trigger function.

Patch-set: 2
Label: Code-Review=0