Update patch set 1
Patch Set 1:

CL:225771 fixed multi-character input in STOP mode. I'm not sure if this was the problem that you were seeing though. Basically the fix was to just wake the core earlier so that we can fetch the first character in time before it's overrun by the subsequent characters. This is of course a race condition, but I didn't find a better way to do it.

What was the problem if you enable Rx DMA with STOP mode?

Patch-set: 1