gesture: double tap for battery accelereomter ODR
Double the output data rate setting for accels for tap
for battery because with the lsm6ds0 sensor, the current
ODR does not guarantee that we get fresh data every time
we do a read.

BUG=chrome-os-partner:34592
BRANCH=samus
TEST=load onto samus, go to G3, use tapinfo on to see
ODR set to 476000Hz. add printf to lsm6ds0 to print when
data is not ready, and see that without this change, the
printf is hit once a minute or so, and with this change,
the printf is never hit, meaning new data is always ready.

Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237665
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
(cherry picked from commit 9d25fa90061fecbbd34ab9be57be424ad1da1498)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: Ib09cba0d70642f59df8a35b5b6bb4fd7b9a4fa3d