pd: implement new type-C connect state machine
Implement the new type-C connect state machine which removes
lock and hold times and adds a debounce time to make sure
CC lines settle before going into the attached state.

This also adds detection of accessories, but doesn't do anything
when an accessory is detected.

WARNING: the new type-C connect state machine is not compatible
with the old state machine. a source with the new state machine
plugged into a drp with the old state machine will NEVER make a
connection because the sink will keep dual-role toggling while the
source is trying to debounce the CC line.

BUG=chrome-os-partner:33680
BRANCH=samus
TEST=test samus connected zinger and samus connected to samus. make
sure that the connection is always formed. also, took twinkie
capture and noticed that with the new sink debounce time, the sink
always fails to respond to the first set of source cap packets, but
then responds to second set 100ms later.

Change-Id: I91a7a6031bc35082cc19d7697142e4aa92ef46f2
Signed-off-by: Alec Berg <alecaberg@chromium.org>