samus: pd: reject power swap if not in dual-role mode
Samus should reject request for a power swap if it is not in a
dual-role mode, for instance if in S3 or lower.

BUG=chrome-os-partner:35141
BRANCH=samus
TEST=load on two samus'. turn on samus off, then request a power
swap from the samus that is on and see that it is rejected.

Change-Id: Ifb23febac967c981251c3932856e1e942818694f
Signed-off-by: Alec Berg <alecaberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238963
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
(cherry picked from commit 55c3456b52e7ed6a8f35c1dec3b5b0036239e8eb)
Signed-off-by: Alec Berg <alecaberg@chromium.org>