Update patch set 2
Patch Set 2:

Because our policy in general is not to provide power in S5, and only provide power in S3 if we were providing power on that port when we entered S3.

If we want to change that policy, then the drp_state would remain in the TOGGLE_ON position and this change would work just as well.

The problem is with DRP toggling off, if you attempt to swap power then you get stuck in the opposite role from what you wanted to be. For example if you have toggling off and you are a sink, and you attempt a swap, you will end up as a source with no way to get back to being a sink.

The whole point of the drp_state variable is to say whether we are currently a dual-role device or not. If we aren't currently acting as a dual-role device, then we shouldn't accept a power swap.

Patch-set: 2