Update patch set 1
Patch Set 1:

I think this should be a CONFIG option at a minimum, but it may be even better to have it default to on but allow a host command to disable it.  This way if people are not running Chrome OS kernel/userland they will still get lightbar transitions on suspend/resume.

Patch-set: 1
Label: Code-Review=0