samus: pd: fix potential junk at end of tx transmission
Fix potential junk at end of PD TX transmit by adding to the DMA
transmit complete interrupt a blocking wait for SPI to finish and
then immediately disable SPI clock. This means we block in an
interrupt function for approximately 45us at the end of every transmit.
But, this is the highest priority thing going on anyway.

Note, there is still a potential for junk if both ports are transmitting
at the same time and finish very close to the same time.

BUG=chrome-os-partner:34600
BRANCH=samus
TEST=load onto samus and test communications with zinger. tested specifically
with an old zinger CL, https://chromium-review.googlesource.com/#/c/226118/11,
which watchdogs when samus has junk at end of transmit. Tested without this
CL and verified we could never successfully flash zinger over PD due to this
watchdog and verified on scope presence of junk. Then tested with this change
and was able to successfully flash zinger.

Change-Id: If0cd9ab0551d36a7d7dc10232b6476dd56735972
Signed-off-by: Alec Berg <alecaberg@chromium.org>