samus: Require 2% battery charge for power-on
If the battery isn't charged to 2%, don't allow Samus to power-on,
regardless of whether a charger is currently plugged. This threshold was
previously 1%, but it must be increased to allow power manager more time
to do a clean shutdown.

Note that the lightbar pulses red while AC is plugged and charge is
inhibited, and the limit only takes effect when WP is active.

BUG=chrome-os-partner:35188
TEST=Manual on Samus. Deplete battery to 1%, verify that power button
press is inhibited and that lightbar pulses red. Charge to 2% and verify
system powers on.
BRANCH=Samus

Change-Id: I57658cab677c0c7363a8a482e89999ba1c86c607
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>