Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Any CONFIGs you define have to be in include/config.h as per:
http://www.chromium.org/chromium-os/ec-development#TOC-Coding-style

Why do you need this driver on the EC? It's really complicated for it to work like that. So it would be wouldn't it be better if the AP dealt with this?

Patch-set: 2
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1