Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

This seems fine to me.  It's what we do on most other ECs already; it's just stm32f0 which was missing the check.

Might want to use the same implementation as the other chips (check sysjump flag), but that's really a minor issue which only causes slightly different behavior if you're in RW and 'sysjump ro'.  The code in this patch would try to apply the different write protection setting and reboot; other chips wouldn't.  Potentially affects flashrom, but only in this weird corner case.

Patch-set: 1
Label: Code-Review=+1