Update patch set 1
Patch Set 1: Commit-Queue+1

Tested on DVT board and I am able to boot and install an image w/o a heatsink.  W/o a heatsink I regularly see temps >100C but PROCHOT is asserted and that manages to keep it <104C so it does not shut down.

Patch-set: 1
Label: Commit-Queue=+1