samus: Move delay for booting without a battery
In the samus power sequence steps there is a delay of 500ms when
booting without a battery attached.  This delay is currently set
to happen in the S5->S3 transition time, which seems to confuse
sequencing on some boards as we have already set DPWROK and
RSMRST and so this 500ms delay can mess up timing.

Move this delay to happen after 5V is up but before any of the
triggers to the chipset to start sequencing are set.

BUG=chrome-os-partner:31583
BRANCH=samus
TEST=boot on samus pre-DVT and DVT without a battery

Change-Id: I9519c336b121c86183c8368282f6c94b32987c16
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241180
Reviewed-by: Alec Berg <alecaberg@chromium.org>
(cherry picked from commit 0747b700fa4ee7ee2a22560d76f5eb2afb822591)
Signed-off-by: Alec Berg <alecaberg@chromium.org>