Update patch set 1
Patch Set 1: Code-Review-1

This is actually a very useful command. I use it a lot when validating power consumption to see if my change significantly changes the fraction of time we spend in deep sleep.

The strings are quite verbose though, so I'm not opposed to shortening the strings.

Patch-set: 1
Label: Code-Review=-1