Update patch set 1
Patch Set 1: Verified+1

Not sure this is the best way to do it, but this fixes the problem of booting without battery (with software sync enabled) on samus.

It makes sense that this should hold for all boards, the only reason this is coming up now is because samus is the first board with Raiden where the input current limit is really variable.

Patch-set: 1
Label: Verified=+1