Update patch set 2
Patch Set 2:

Yes, this will fix the current problem, but it will break not allowing boot with dead battery because this check happens before we know our battery percentage. Seems like we have to either sample the battery earlier, or move this check until later. It probably used to work when we did it in the S5 -> S3 transition, but that had other problems.

Patch-set: 2